requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.1
pyyaml==6.0.1
plotly==5.24.1
streamlit==1.38.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
great_expectations==0.18.14
mlflow==2.12.1
evidently==0.4.25
pytest==8.3.3
openpyxl>=3.1.0

